shine bright

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,20 +5,20 @@
  */
 int main(void)
 {
-	unsigned long int a = 1;
-	unsigned long int b = 2;
-	unsigned long int b1 = 0;
+	long double int a = 1;
+	long double int b = 2;
+	long double int b1 = 0;
 	int loop = 0;
 
 	while (loop < 97)
 	{
-		printf("%lu, ", a);
+		printf("%Lf, ", a);
 		b1 = b;
 		b = a + b;
 		a = b1;
 		loop++;
 	}
-	printf("%lu\n", a);
+	printf("%Lf\n", a);
 	return (0);
 }
